fix: move temporary Gravity error handler to a safe location
