+++ src/lib/gravityErrorHandler.ts
@@ -36,27 +36,6 @@ export const GravityMutationErrorType = new GraphQLObjectType<
   }),
 })
 
-export const formatGravityHttpError = (error) => {
-  const errorBody =
-    typeof error.body === "string" ? JSON.parse(error.body) : error.body
-  const errorType = errorBody?.type || "error"
-  const message = errorBody?.message || errorBody?.error
-  // confusing part: gravity returns field errors in the body.detail, but in some cases (PaymentError for instance),
-  // body.detail includes a string message. GravityMutationErrorType#fieldErrors hold the actual field errors
-  // and GravityMutationErrorType#detail - the string message
-  const fieldErrors = formatGravityErrorDetails(errorBody?.detail || {})
-  const detail =
-    typeof errorBody?.detail === "object" ? null : errorBody?.detail
-
-  return {
-    type: errorType,
-    statusCode: error.statusCode,
-    message,
-    fieldErrors,
-    detail,
-  }
-}
-
 export const formatGravityError = (error) => {
   const errorSplit = error.message?.split(" - ")
 
@@ -144,7 +123,7 @@ type FieldErrorType = {
   message: string
 }
 
-const formatGravityErrorDetails = (
+export const formatGravityErrorDetails = (
   detail: Record<string, string[]>
 ): FieldErrorType[] => {
   const fieldErrors: FieldErrorType[] = []
