+++ src/schema/v2/me/createCollectionMutation.ts
@@ -6,8 +6,8 @@ import {
 } from "graphql"
 import { mutationWithClientMutationId } from "graphql-relay"
 import {
-  formatGravityHttpError,
   GravityMutationErrorType,
+  formatGravityErrorDetails,
 } from "lib/gravityErrorHandler"
 import { ResolverContext } from "types/graphql"
 import { CollectionType } from "./collection"
@@ -87,3 +87,24 @@ export const createCollectionMutation = mutationWithClientMutationId<
     }
   },
 })
+
+const formatGravityHttpError = (error) => {
+  const errorBody =
+    typeof error.body === "string" ? JSON.parse(error.body) : error.body
+  const errorType = errorBody?.type || "error"
+  const message = errorBody?.message || errorBody?.error
+  // confusing part: gravity returns field errors in the body.detail, but in some cases (PaymentError for instance),
+  // body.detail includes a string message. GravityMutationErrorType#fieldErrors hold the actual field errors
+  // and GravityMutationErrorType#detail - the string message
+  const fieldErrors = formatGravityErrorDetails(errorBody?.detail || {})
+  const detail =
+    typeof errorBody?.detail === "object" ? null : errorBody?.detail
+
+  return {
+    type: errorType,
+    statusCode: error.statusCode,
+    message,
+    fieldErrors,
+    detail,
+  }
+}
